lib: additional performance improvements to checkout

+++ lib/src/local_working_copy.rs
@@ -544,6 +544,7 @@ fn sparse_patterns_from_proto(
 fn create_parent_dirs(
     working_copy_path: &Path,
     repo_path: &RepoPath,
+    reserved_path_cache: &mut HashSet<PathBuf>,
 ) -> Result<Option<PathBuf>, CheckoutError> {
     let (parent_path, basename) = repo_path.split().expect("repo path shouldn't be root");
     let mut dir_path = working_copy_path.to_owned();
@@ -574,7 +575,7 @@ fn create_parent_dirs(
         };
         // Invalid component (e.g. "..") should have been rejected.
         // The current dir_path should be an entry of dir_path.parent().
-        reject_reserved_existing_path(&dir_path).inspect_err(|_| {
+        reject_reserved_existing_path(&dir_path, reserved_path_cache).inspect_err(|_| {
             if new_dir_created {
                 fs::remove_dir(&dir_path).ok();
             }
@@ -596,7 +597,8 @@ fn create_parent_dirs(
 /// If the existing file points to ".git" or ".jj", this function returns an
 /// error.
 fn remove_old_file(disk_path: &Path) -> Result<bool, CheckoutError> {
-    reject_reserved_existing_path(disk_path)?;
+    let mut empty_cache = HashSet::new();
+    reject_reserved_existing_path(disk_path, &mut empty_cache)?;
     match fs::remove_file(disk_path) {
         Ok(()) => Ok(true),
         Err(err) if err.kind() == io::ErrorKind::NotFound => Ok(false),
@@ -618,7 +620,11 @@ fn remove_old_file(disk_path: &Path) -> Result<bool, CheckoutError> {
 /// error.
 ///
 /// This function can fail if `disk_path.parent()` isn't a directory.
-fn can_create_new_file(disk_path: &Path) -> Result<bool, CheckoutError> {
+fn can_create_new_file(
+    delete_file: bool,
+    disk_path: &Path,
+    reserved_path_cache: &mut HashSet<PathBuf>,
+) -> Result<(bool, Option<File>), CheckoutError> {
     // New file or symlink will be created by caller. If it were pointed to by
     // name ".git" or ".jj", git/jj CLI could be tricked to load configuration
     // from an attacker-controlled location. So we first test the path by
@@ -642,23 +648,37 @@ fn can_create_new_file(disk_path: &Path) -> Result<bool, CheckoutError> {
         },
     };
 
-    let new_file_created = new_file.is_some();
-
     if let Some(new_file) = new_file {
-        reject_reserved_existing_file(new_file, disk_path).inspect_err(|_| {
-            // We keep the error from `reject_reserved_existing_file`
-            let _ = fs::remove_file(disk_path);
-        })?;
+        // Try to clone this descriptor. If it fails, the caller will account for it.
+        let mut return_file = new_file.try_clone().ok();
 
-        fs::remove_file(disk_path).map_err(|err| CheckoutError::Other {
-            message: format!("Failed to remove temporary file {}", disk_path.display()),
-            err: err.into(),
-        })?;
+        let reject_result = reject_reserved_existing_file(new_file, disk_path, reserved_path_cache);
+        if reject_result.is_err() || delete_file {
+            // Ensure there are no more references to this file before
+            // we attempt to remove it
+            let _ = return_file.take();
+
+            // Try to remove this file.
+            let remove_file_result = fs::remove_file(disk_path);
+
+            // The `fs::remove_file` result doesn't matter if `reject_reserved*` failed.
+            // In that case, we bubble up the `reject_result` error.
+            if reject_result.is_err() {
+                reject_result?;
+            } else {
+                remove_file_result.map_err(|err| CheckoutError::Other {
+                    message: format!("Failed to remove temporary file {}", disk_path.display()),
+                    err: err.into(),
+                })?;
+            }
+        }
+
+        Ok((true, return_file))
     } else {
-        reject_reserved_existing_path(disk_path)?;
-    }
+        reject_reserved_existing_path(disk_path, reserved_path_cache)?;
 
-    Ok(new_file_created)
+        Ok((false, None))
+    }
 }
 
 const RESERVED_DIR_NAMES: &[&str] = &[".git", ".jj"];
@@ -688,13 +708,21 @@ fn map_samefile_result(
 /// file descriptor.
 ///
 /// See [`reject_reserved_existing_handle`] for more info.
-fn reject_reserved_existing_file(file: File, disk_path: &Path) -> Result<(), CheckoutError> {
+fn reject_reserved_existing_file(
+    file: File,
+    disk_path: &Path,
+    reserved_path_cache: &mut HashSet<PathBuf>,
+) -> Result<(), CheckoutError> {
+    if can_skip_reject_reserved(disk_path, reserved_path_cache) {
+        return Ok(());
+    }
+
     let Some(file_handle) = map_samefile_result(disk_path, same_file::Handle::from_file(file))?
     else {
         return Ok(());
     };
 
-    reject_reserved_existing_handle(file_handle, disk_path)
+    reject_reserved_existing_handle(file_handle, disk_path, reserved_path_cache)
 }
 
 /// Wrapper for [`reject_reserved_existing_handle`] which converts
@@ -706,14 +734,33 @@ fn reject_reserved_existing_file(file: File, disk_path: &Path) -> Result<(), Che
 ///
 /// Incurs an additional syscall cost to open and close the file
 /// descriptor/`HANDLE` for `disk_path`.
-fn reject_reserved_existing_path(disk_path: &Path) -> Result<(), CheckoutError> {
+fn reject_reserved_existing_path(
+    disk_path: &Path,
+    reserved_path_cache: &mut HashSet<PathBuf>,
+) -> Result<(), CheckoutError> {
+    if can_skip_reject_reserved(disk_path, reserved_path_cache) {
+        return Ok(());
+    }
+
     let Some(disk_handle) =
         map_samefile_result(disk_path, same_file::Handle::from_path(disk_path))?
     else {
         return Ok(());
     };
 
-    reject_reserved_existing_handle(disk_handle, disk_path)
+    reject_reserved_existing_handle(disk_handle, disk_path, reserved_path_cache)
+}
+
+/// Check if we can skip the `reject_reserved_existing*` check based on
+/// the reserved paths are in the provided `reserved_path_cache`.
+///
+/// If all items are present, that means that none of those directories exist.
+fn can_skip_reject_reserved(disk_path: &Path, reserved_path_cache: &HashSet<PathBuf>) -> bool {
+    let parent_dir_path = disk_path.parent().expect("content path shouldn't be root");
+    RESERVED_DIR_NAMES.iter().all(|name| {
+        let reserved_path = parent_dir_path.join(name);
+        reserved_path_cache.contains(&reserved_path)
+    })
 }
 
 /// Suppose the `disk_path` exists, checks if the last component points to
@@ -726,13 +773,19 @@ fn reject_reserved_existing_path(disk_path: &Path) -> Result<(), CheckoutError> 
 fn reject_reserved_existing_handle(
     disk_handle: same_file::Handle,
     disk_path: &Path,
+    reserved_path_cache: &mut HashSet<PathBuf>,
 ) -> Result<(), CheckoutError> {
     let parent_dir_path = disk_path.parent().expect("content path shouldn't be root");
     for name in RESERVED_DIR_NAMES {
         let reserved_path = parent_dir_path.join(name);
+        if reserved_path_cache.contains(&reserved_path) {
+            continue;
+        }
+
         let Some(reserved_handle) =
-            map_samefile_result(disk_path, same_file::Handle::from_path(reserved_path))?
+            map_samefile_result(disk_path, same_file::Handle::from_path(&reserved_path))?
         else {
+            let _ = reserved_path_cache.insert(reserved_path);
             continue;
         };
 
@@ -1765,14 +1818,14 @@ impl FileSnapshotter<'_> {
 
 /// Functions to update local-disk files from the store.
 impl TreeState {
-    async fn write_file(
+    async fn write_file_with_path(
         &self,
         disk_path: &Path,
         contents: impl AsyncRead + Send + Unpin,
         executable: bool,
         apply_eol_conversion: bool,
     ) -> Result<FileState, CheckoutError> {
-        let mut file = File::options()
+        let file = File::options()
             .write(true)
             .create_new(true) // Don't overwrite un-ignored file. Don't follow symlink.
             .open(disk_path)
@@ -1780,6 +1833,19 @@ impl TreeState {
                 message: format!("Failed to open file {} for writing", disk_path.display()),
                 err: err.into(),
             })?;
+
+        self.write_file(file, disk_path, contents, executable, apply_eol_conversion)
+            .await
+    }
+
+    async fn write_file(
+        &self,
+        mut file: File,
+        disk_path: &Path,
+        contents: impl AsyncRead + Send + Unpin,
+        executable: bool,
+        apply_eol_conversion: bool,
+    ) -> Result<FileState, CheckoutError> {
         let contents = if apply_eol_conversion {
             self.target_eol_strategy
                 .convert_eol_for_update(contents)
@@ -1827,8 +1893,28 @@ impl TreeState {
         Ok(FileState::for_symlink(&metadata))
     }
 
+    async fn write_conflict_with_path(
+        &self,
+        disk_path: &Path,
+        contents: &[u8],
+        executable: bool,
+    ) -> Result<FileState, CheckoutError> {
+        let file = OpenOptions::new()
+            .write(true)
+            .create_new(true) // Don't overwrite un-ignored file. Don't follow symlink.
+            .open(disk_path)
+            .map_err(|err| CheckoutError::Other {
+                message: format!("Failed to open file {} for writing", disk_path.display()),
+                err: err.into(),
+            })?;
+
+        self.write_conflict(file, disk_path, contents, executable)
+            .await
+    }
+
     async fn write_conflict(
         &self,
+        mut file: File,
         disk_path: &Path,
         contents: &[u8],
         executable: bool,
@@ -1839,14 +1925,6 @@ impl TreeState {
             .await
             .map_err(|err| CheckoutError::Other {
                 message: "Failed to convert the EOL when writing a merge conflict".to_string(),
-                err: err.into(),
-            })?;
-        let mut file = OpenOptions::new()
-            .write(true)
-            .create_new(true) // Don't overwrite un-ignored file. Don't follow symlink.
-            .open(disk_path)
-            .map_err(|err| CheckoutError::Other {
-                message: format!("Failed to open file {} for writing", disk_path.display()),
                 err: err.into(),
             })?;
         let size = copy_async_to_sync(contents, &mut file)
@@ -1949,6 +2027,7 @@ impl TreeState {
 
         let mut prev_created_path: Option<PathBuf> = None;
         let mut path_prefix = PathBuf::new();
+        let mut reserved_path_cache = HashSet::new();
 
         while let Some((path, data)) = diff_stream.next().await {
             let (before, after) = data?;
@@ -2032,8 +2111,11 @@ impl TreeState {
                 // Cache this path for the next iteration.
                 prev_created_path = Some(this_path);
 
-                let Some(disk_path) =
-                    create_parent_dirs(adjusted_working_copy_path, adjusted_diff_file_path)?
+                let Some(disk_path) = create_parent_dirs(
+                    adjusted_working_copy_path,
+                    adjusted_diff_file_path,
+                    &mut reserved_path_cache,
+                )?
                 else {
                     changed_file_states.push((path, FileState::placeholder()));
                     stats.skipped_files += 1;
@@ -2045,11 +2127,40 @@ impl TreeState {
 
             // If the path was present, check reserved path first and delete it.
             let present_file_deleted = before.is_present() && remove_old_file(&disk_path)?;
+
+            let mut new_file = None;
+
             // If not, create temporary file to test the path validity.
-            if !present_file_deleted && !can_create_new_file(&disk_path)? {
-                changed_file_states.push((path, FileState::placeholder()));
-                stats.skipped_files += 1;
-                continue;
+            if !present_file_deleted {
+                let delete_file = match &after {
+                    // These operations can re-use the file returned by `can_create_new_file`
+                    MaterializedTreeValue::File(_)
+                    | MaterializedTreeValue::FileConflict(_)
+                    | MaterializedTreeValue::OtherConflict { .. } => false,
+
+                    // If it's a symlink AND we have symlink support, delete it.
+                    MaterializedTreeValue::Symlink { .. } if self.symlink_support => true,
+
+                    // These operations either don't use the new file, or might
+                    // try to do diverse operations that _may_ expect no file to
+                    // be present (like creating a symlink)
+                    MaterializedTreeValue::Absent
+                    | MaterializedTreeValue::AccessDenied(_)
+                    | MaterializedTreeValue::Symlink { .. }
+                    | MaterializedTreeValue::GitSubmodule(_)
+                    | MaterializedTreeValue::Tree(_) => true,
+                };
+
+                let (created, temp_new_file) =
+                    can_create_new_file(delete_file, &disk_path, &mut reserved_path_cache)?;
+                new_file = temp_new_file;
+
+                if !created {
+                    changed_file_states.push((path, FileState::placeholder()));
+                    stats.skipped_files += 1;
+
+                    continue;
+                }
             }
 
             // TODO: Check that the file has not changed before overwriting/removing it.
@@ -2072,14 +2183,22 @@ impl TreeState {
                     continue;
                 }
                 MaterializedTreeValue::File(file) => {
-                    self.write_file(&disk_path, file.reader, file.executable, true)
-                        .await?
+                    if let Some(new_file) = new_file {
+                        self.write_file(new_file, &disk_path, file.reader, file.executable, true)
+                            .await?
+                    } else {
+                        self.write_file_with_path(&disk_path, file.reader, file.executable, true)
+                            .await?
+                    }
                 }
                 MaterializedTreeValue::Symlink { id: _, target } => {
                     if self.symlink_support {
                         self.write_symlink(&disk_path, target)?
+                    } else if let Some(new_file) = new_file {
+                        self.write_file(new_file, &disk_path, target.as_bytes(), false, false)
+                            .await?
                     } else {
-                        self.write_file(&disk_path, target.as_bytes(), false, false)
+                        self.write_file_with_path(&disk_path, target.as_bytes(), false, false)
                             .await?
                     }
                 }
@@ -2099,9 +2218,23 @@ impl TreeState {
                         merge: self.store.merge_options().clone(),
                     };
                     let contents = materialize_merge_result_to_bytes(&file.contents, &options);
-                    let mut file_state = self
-                        .write_conflict(&disk_path, &contents, file.executable.unwrap_or(false))
-                        .await?;
+
+                    let mut file_state = if let Some(new_file) = new_file {
+                        self.write_conflict(
+                            new_file,
+                            &disk_path,
+                            &contents,
+                            file.executable.unwrap_or(false),
+                        )
+                        .await?
+                    } else {
+                        self.write_conflict_with_path(
+                            &disk_path,
+                            &contents,
+                            file.executable.unwrap_or(false),
+                        )
+                        .await?
+                    };
                     file_state.materialized_conflict_data = Some(MaterializedConflictData {
                         conflict_marker_len: conflict_marker_len.try_into().unwrap_or(u32::MAX),
                     });
@@ -2112,10 +2245,16 @@ impl TreeState {
                     // better than trying to describe the merge.
                     let contents = id.describe();
                     let executable = false;
-                    self.write_conflict(&disk_path, contents.as_bytes(), executable)
-                        .await?
+                    if let Some(new_file) = new_file {
+                        self.write_conflict(new_file, &disk_path, contents.as_bytes(), executable)
+                            .await?
+                    } else {
+                        self.write_conflict_with_path(&disk_path, contents.as_bytes(), executable)
+                            .await?
+                    }
                 }
             };
+
             changed_file_states.push((path, file_state));
         }
         self.file_states
